refactor: :core:ui モジュールについて、ui -> core.ui に変更した。
